Fix version to string conversion crash (wooga/atlas-build-unity#14)
> Note Original PR: https://github.com/wooga/atlas-build-unity/pull/14

Description
===========

The `project.version` property is by default an `Object`. In most cases it will 
be a `String` value. I changed the way we set this `version` value from a direct
assignment into a convention mapping (#9). What is important to note is that
instead of the original `getter` the mapped `closure` will be executed instead.
I added a lot of conversion methods to convert an object into a `String`.
This wasn't done in the mapping. 

I added a very quick regression test in `UnityBuildPluginIntegrationSpec`.
I wanted to do the same for the iOS plugin but this is impossible at the
moment due to the untestable state of the plugin. I also moved some
test/utility methods around to test/execute the logic without major code duplications

Changes
=======

![FIX] `project.version` toString conversion