Improve ArchiveTask provide optional team id (wooga/atlas-build-unity#56)
> Note Original PR: https://github.com/wooga/atlas-build-unity/pull/56

Description
===========

During archive and the resulting code sign, it can happen that Xcode has
issues with ambigious certificate names. It is possible to provide a
`DEVELOPMENT_TEAM` flag during invocation and providing the teamId for
code sign.

This patch adds a new optional task input variable `teamId` to the
`Archive` task and sets the `DEVELOPMENT_TEAM` option during execution
if the property is set. The plugin will provide a default value from the
plugin extension.

Changes
=======

* ![IMPROVE] ![IOS] `ArchiveTask` provide optional team id