Add conditional reset of the keychains before adding build keychain (wooga/atlas-build-unity#72)
> Note Original PR: https://github.com/wooga/atlas-build-unity/pull/72

Description
===========

This patch adds a new action `reset` to the `ListKeychainTask`. This
action is conditional and bount to the environment variable
`ATLAS_BUILD_UNITY_IOS_RESET_KEYCHAINS`. The variable is checked
directly in the task implementation for now.
The plugin configures a new task called `resetKeychains` and adds it as
a dependency to the `addKeychain` task.

This change should help with _ghost keychains_ on the CI system.

Changes
=======

* ![ADD] conditional reset of keychains