Use new fastlane plugin (wooga/atlas-build-unity#89)
> Note Original PR: https://github.com/wooga/atlas-build-unity/pull/89

* Add PilotUpload task

Description
===========

This is the second patch which will soon replace the old `fastlane`
based task from `net.wooga.build-unity-ios`. The new task type
`PilotUpdate` will soon replace the old `PublishTestFlight`
implementation. This new task has nearly the same interface. It uses
like all oter new types the Provider API to get rid of convention
mapping as much as possible. The implementation uses a lot of the
underlying abstract base class `AbstractFastlaneTask` which was
introduced in the last patch. This makes this change rather small. It
supports logfile creation and the ability to add additional arguments in
case fastlane adds more commandline arguments or we need to use a few of
the ones I did not yet implement because of a good reason to have them
in.

Changes
=======

* ![ADD] `PilotUpload` task type

* Add fastlane gradle plugin

Description
===========

This patch adds a simple plugin definiton for `fastlane`. This allows to
setup shared configuration in a single extention. I implemented only two
settings in the provided fastlane extension: `username` and `password`.

| property              | gradle property name              | environment variable                |
| --------------------- | --------------------------------- | ----------------------------------- |
| username              | `fastlane.username`               | `FASTLANE_USERNAME`                 |
| password              | `fastlane.password`               | `FASTLANE_PASSWORD`                 |

In this case, we use the same env names as fastlanes would use as well.

Changes
=======

* ![ADD] ![GRADLE] `FastlanePlugin` with defaults

* Use new fastlane plugin

Description
===========

This patch removes the old task types `ImportProvisioningProfile` and
`PublishTestFlight` and instead uses the newly created counterparts
`SighRenew` and `PilotUpload`. The new tasks work internally the same.

Changes
=======

* ![IMPROVE] use new fastlane tasktypes
* ![REMOVE] old tasktype `ImportProvisionProfile`
* ![REMOVE] old tasktype `PublishTestFlight`