Improve ImportCodeSigningIdentity task (wooga/atlas-build-unity#129)
> Note Original PR: https://github.com/wooga/atlas-build-unity/pull/129

Description
===========

This is a mix of improvment and fix for the `ImportCodeSigningIdentity` task.
The fix is actually located in the `FindIdentityResult` class.
The identities reported by the `secrets` tool when fetched from the keychain
will end with an id value which is normally not used when setting up
signing identities in xcode. I fixed it by always adding the identity with and
without this last id in the result list. I did this so I can still check
for complete identity strings and not use `startsWith` which could lead
to false positives.

The general improvement of the task is the general usability. Since `p12` files
can contain one or more items to import, I decided to reflect that also in the task type
name and API. The name of the type changed to `ImportCodeSigningIdentities`.
Means one can import multiple identities in one call.
Further is the list of to be validated identities optional. So the task
will simply import the `p12` if no signging identities are provided.
The `p12` file is also now kind of optional. The task will skip with `[NO-SOURCE]
when the file is not set.

I also added a new property to control the access rights from other
binaries to the imported items in the keychain. The main usecase is to
allow `usr/bin/codesign` access. This has the drawback that I can't skip
the task even if I find the identity in the keychain. The reason is
that the `security` tool has no way of fetching the access list for
a given item. So the task has to assume the provided path or multiple paths
are not set and needs to reimport the item/items.

I pulled in the latest version of `com.wooga.gradle:gralde-commons`
and `com.wooga.gradle:gradle-commons-test`. I adjusted the testsetup
to use the utility methods from these libraries. I did this only in the
test for the `ImportCodeSigningIdentities` class.

Changes
=======

* ![IMPROVE] `ImportCodeSigningIdentities` task type
* ![IMPROVE] signing identites matching logic in `FindIdentityResult`
* ![RENAME] `ImportCodeSigningIdentity` to `ImportCodeSigningIdentities`
* ![ADD] `applicationAccessPaths` property in `ImportCodeSigningIdentities`
* ![IMPROVE] `ImportCodeSigningIdentities` usability by making more properties optional