Add support for wildcard profile fetch
Description
===========

The plugin assumes that the bundle id of the app and the app id
for the profile is one and the same. That is generally true when
dealing with normal build setups. In case of wildcard app ids and
the profiles we need to treat the fetching differently.
I adjusted the setter for the profiles to be fetched with fastlane
in such a case that it will check if the extension property
`provisioningProfileAppId` is set and ends with `.*`. This indicates
that you actually want a wildcard profile fetched. The map of profiles
to fetch which is normally pulled from the `exportOptions.plist` file is
adjusted so that it will replace all wildcard matching bundle ids
with the `provisioningProfileAppId`.

There are many edge cases not covered here. And I also think that we could
maybe simplify this setup by breaking up the difference between app id of the
bundle and app id which is setup in the developer console. The terms are used
interchangeable in most documentations because in 90% of the cases they
refer to the same value.

Changes
=======

* ![ADD] support for wildcard profile fetch