Add artifactory repository handling to pod install task
Description
===========

In order to handle custom artifactory repositories for cocoapods
I added a new property to the `PodInstall` task.
When the property `artRepositories` is set with a list of
repository URL's, the task will call `pod art-repo add ...` and
`pod art-repo update` for each configured repository URL.
It is assumed that the art-cocoapods extension is installed.

Changes
=======

* ![ADD] artifactory repository handling in `PodInstall` task